pyproject.toml
[project]
name = "chainlit-bootstrap"
version = "0.1.0"
description = "A Chainlit starter application with voice input and document QA"
requires-python = ">=3.12,<3.14"
authors = [
    {name = "Chainlit Starter", email = "starter@example.com"}
]

dependencies = [
    "chainlit>=2.9.0",
    "chromadb>=1.3.4",
    "matplotlib>=3.9.0",
    "seaborn>=0.13.0",
    "openai>=2.8.0",
    "sqlalchemy>=2.0.44",
    "aiosqlite>=0.21.0",
    "asyncpg>=0.30.0",
    "llama-index-core>=0.10.0",
    "llama-index-llms-openai>=0.1.0",
    "llama-index-embeddings-openai>=0.1.0",
    "llama-index-vector-stores-chroma>=0.1.0",
    "tiktoken>=0.12.0",
    "tavily-python>=0.3.8",
    "typing_extensions>=4.12.0",
    "python-dotenv>=1.2.1",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.14.5",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
]
parlant = [
    "parlant>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["chainlit_bootstrap"]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by formatter
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint.isort]
known-first-party = ["chainlit", "chainlit_bootstrap"]

